backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic==0.49.0
pydantic>=2.0
python-multipart==0.0.20
httpx==0.28.1
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-analytics-data>=0.18.0
apscheduler>=3.10.0
google-ads>=29.0.0
facebook-business>=24.0.0
